vessels activity in vessel group report component
