+++ apps/fishing-map/features/reports/activity/ReportActivityEvolution.tsx
@@ -15,10 +15,9 @@ import { DateTime } from 'luxon'
 import { FourwingsInterval } from '@globalfishingwatch/deck-loaders'
 import i18n from 'features/i18n/i18n'
 import { formatDateForInterval, getUTCDateTime } from 'utils/dates'
-import { toFixed } from 'utils/shared'
 import { formatI18nNumber } from 'features/i18n/i18nNumber'
-import { ReportActivityProps } from 'features/reports/activity/ReportActivity'
 import { formatEvolutionData } from 'features/reports/areas/reports-timeseries.utils'
+import { ReportActivityProps } from 'features/reports/activity/ReportActivityGraph'
 import { tickFormatter } from '../areas/reports.utils'
 import styles from './ReportActivityEvolution.module.css'
