+++ apps/fishing-map/features/reports/activity/vessels/ReportVessels.tsx
@@ -1,15 +1,15 @@
 import React, { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
-import ReportVesselsGraphSelector from 'features/reports/areas/vessels/ReportVesselsGraphSelector'
+import ReportVesselsGraphSelector from 'features/reports/activity/vessels/ReportVesselsGraphSelector'
 import { selectActiveReportDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { selectReportCategory } from 'features/app/selectors/app.reports.selector'
 import ReportSummaryTags from 'features/reports/areas/summary/ReportSummaryTags'
 import { FIELDS, getCommonProperties } from 'features/reports/areas/reports.utils'
 import { PROPERTIES_EXCLUDED } from 'features/reports/areas/summary/ReportSummary'
-import { ReportActivityUnit } from '../Report'
-import { selectReportVesselFilter } from '../reports.config.selectors'
-import { ReportCategory } from '../reports.types'
+import { ReportActivityUnit } from 'features/reports/areas/Report'
+import { selectReportVesselFilter } from 'features/reports/areas/reports.config.selectors'
+import { ReportCategory } from 'features/reports/areas/reports.types'
 import ReportVesselsGraph from './ReportVesselsGraph'
 import ReportVesselsFilter from './ReportVesselsFilter'
 import ReportVesselsTable from './ReportVesselsTable'
@@ -17,7 +17,7 @@ import styles from './ReportVessels.module.css'
 
 type ReportVesselTableProps = {
   activityUnit: ReportActivityUnit
-  reportName: string
+  reportName?: string
 }
 
 export default function ReportVessels({ activityUnit, reportName }: ReportVesselTableProps) {
