+++ apps/fishing-map/features/reports/activity/vessels/ReportVesselsGraph.module.css
