+++ apps/fishing-map/features/reports/activity/vessels/ReportVesselsTableFooter.module.css
