+++ apps/fishing-map/features/reports/vessel-groups/vessels/vessel-group-report-vessels.selectors.ts
@@ -9,7 +9,6 @@ import {
   selectVGRVesselPage,
 } from 'features/reports/vessel-groups/vessel-group.config.selectors'
 import { formatInfoField, getVesselGearTypeLabel, getVesselShipTypeLabel } from 'utils/info'
-import { cleanFlagState } from 'features/reports/areas/reports.selectors'
 import { t } from 'features/i18n/i18n'
 import {
   FILTER_PROPERTIES,
@@ -21,6 +20,7 @@ import {
   selectVGRVesselsOrderProperty,
   selectVGRVesselsSubsection,
 } from 'features/reports/vessel-groups/vessel-group.config.selectors'
+import { cleanFlagState } from 'features/reports/activity/vessels/report-activity-vessels.utils'
 import { selectVGRVessels } from '../vessel-group-report.slice'
 import { VesselGroupReportVesselParsed } from './vessel-group-report-vessels.types'
